+++ src/main/scala/com/anilsenay/controllers/CategoryController.scala
@@ -18,17 +18,22 @@ class CategoryController(dbService: CategoryService.type) extends BaseController
     } ~
     post {
       pathEndOrSingleSlash {
-      entity(as[Category]) { category =>
-        val saved = dbService.insertCategory(category.categoryName)
-        onComplete(saved) {
-          case Success(updatedRows) => {
-            logger.info(s"Inserted category")
-            complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
-          }
-          case Failure(e) => {
-            logger.error(s"Failed to insert category", e)
-            complete(StatusCodes.InternalServerError)
+        entity(as[Category]) { category =>
+          authenticate { authUser =>
+            if (authUser.isAdmin) {
+              val saved = dbService.insertCategory(category.categoryName)
+              onComplete(saved) {
+                case Success(updatedRows) => {
+                  logger.info(s"Inserted category")
+                  complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
+                }
+                case Failure(e) => {
+                  logger.error(s"Failed to insert category", e)
+                  complete(StatusCodes.InternalServerError)
+                }
+              }
             }
+            else complete(StatusCodes.Unauthorized)
           }
         }
       }
@@ -36,26 +41,36 @@ class CategoryController(dbService: CategoryService.type) extends BaseController
     put {
       path(LongNumber) { id =>
         entity(as[Category]) { category =>
-          val updated = dbService.update(id, category.categoryName)
-          onComplete(updated) {
-            case Success(updatedRows) => complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
-            case Failure(e) => {
-              logger.error(s"Failed to update category ${id}", e)
-              complete(StatusCodes.InternalServerError)
+          authenticate { authUser =>
+            if (authUser.isAdmin) {
+              val updated = dbService.update(id, category.categoryName)
+              onComplete(updated) {
+                case Success(updatedRows) => complete(JsObject("updatedRows" -> JsNumber(updatedRows)))
+                case Failure(e) => {
+                  logger.error(s"Failed to update category ${id}", e)
+                  complete(StatusCodes.InternalServerError)
+                }
+              }
             }
+            else complete(StatusCodes.Unauthorized)
           }
         }
       }
     } ~
     delete {
       path(LongNumber) { id =>
-        val deleted = dbService.delete(id)
-        onComplete(deleted) {
-          case Success(updatedRows) => complete(JsObject("deletedRows" -> JsNumber(updatedRows)))
-          case Failure(e) => {
-            logger.error(s"Failed to delete category ${id}", e)
-            complete(StatusCodes.InternalServerError)
+        authenticate { authUser =>
+          if (authUser.isAdmin) {
+            val deleted = dbService.delete(id)
+            onComplete(deleted) {
+              case Success(updatedRows) => complete(JsObject("deletedRows" -> JsNumber(updatedRows)))
+              case Failure(e) => {
+                logger.error(s"Failed to delete category ${id}", e)
+                complete(StatusCodes.InternalServerError)
+              }
+            }
           }
+          else complete(StatusCodes.Unauthorized)
         }
       }
     }
